requirements.txt
fastapi
uvicorn
python-multipart
chromadb
pypdf
python-docx
httpx
